+++ lib/action_client/submission_job.rb
@@ -1,6 +1,8 @@
 module ActionClient
   class SubmissionJob < ActiveJob::Base
-    class HttpStatusMismatchError < ActionClient::Error
+    class DiscardStatusError < ActionClient::Error
+    end
+    class RetryStatusError < ActionClient::Error
     end
 
     attr_reader :response
@@ -24,7 +26,7 @@ module ActionClient
     end
 
     def self.retry_on(*exception_classes, only_status: nil, except_status: nil, **options, &block)
-      ([HttpStatusMismatchError] + exception_classes).each do |exception_class|
+      ([RetryStatusError] + exception_classes).each do |exception_class|
         super(exception_class, **options, &block)
       end
 
@@ -32,7 +34,21 @@ module ActionClient
         after_perform(only_status: only_status, except_status: except_status) do
           filter = [only_status, except_status].detect(&:present?)
 
-          raise HttpStatusMismatchError, "#{response.status} does not match #{filter}"
+          raise RetryStatusError, "#{response.status} does not match #{filter}"
+        end
+      end
+    end
+
+    def self.discard_on(*exception_classes, only_status: nil, except_status: nil, &block)
+      ([DiscardStatusError] + exception_classes).each do |exception_class|
+        super(exception_class, &block)
+      end
+
+      if [only_status, except_status].compact.any?
+        after_perform(only_status: only_status, except_status: except_status) do
+          filter = [only_status, except_status].detect(&:present?)
+
+          raise DiscardStatusError, "#{response.status} does not match #{filter}"
         end
       end
     end